ltx gemma: fold device-time profiler into test_gemma_full.py
Collapse the three standalone gemma harnesses into one file: the Tracy per-op
profiler of the full encode now lives beside the PCC parity test, skip-gated on
GEMMA_PROF=1 so it can't silently green the default (asserting) suite — it needs
`python -m tracy` to capture anyway. Drops the forward-only subset and the
median-of-N latency tool (perf deltas now read from the profiler's single-shot
host-wall).
